+++ Code/filters.h
@@ -2,11 +2,11 @@
 #define FILTERS_H
 
 // Feel free to change return statement and arguments
-void lowPassFilter(int input[], int output[], int n);
-void highPassFilter(int input[], int output[], int n);
-void derivativeFilter(int input[], int output[], int n);
-void squareFilter(int output[], int n);
-void movingWindowIntegration(int input[], int output[], int n);
+void lowPassFilter(double input[], double output[], int n);
+void highPassFilter(double input[], double output[], int n);
+void derivativeFilter(double input[], double output[], int n);
+void squareFilter(double output[], int n);
+void movingWindowIntegration(double input[], double output[], int n);
 
 // Implement additional filters
